+++ packages/terra-button-group/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 * Changed
-  * Replaced `terra-button` with `terra-icon-button`
+  * Replaced `terra-button` with `IconButton`
 
 ## 3.60.2 - (February 15, 2022)