+++ packages/terra-button/CHANGELOG.md
@@ -2,10 +2,10 @@
 
 ## Unreleased
 * Added
-  * Added new subcomponent `terra-icon-button` to display icon.
+  * Added new subcomponent `IconButton` to display icon.
 
 * Breaking Changes
-  * `icon`, `isIconOnly` and `isReveresed` props have been moved from `terra-button` to `terra-icon-button`.
+  * `icon`, `isIconOnly` and `isReveresed` props have been moved from `terra-button` to `IconButton`.
 
 ## 3.61.0 - (February 15, 2022)